test(device-ts): cover in-flight Whisper stop in the bun suite
Move the #12964 cases into index.test.ts so `bun test` runs them, and
drop the duplicate Node-only file. Fixes #12964.
