+++ sdks/device/typescript/src/stt/index.test.ts
@@ -102,6 +102,90 @@ describe('createWhisperTranscriber', () => {
 
     expect(transcripts).toEqual(['final transcript']);
   });
+
+  test('delivers an in-flight full batch after stop and ignores later audio', async () => {
+    const transcripts: string[] = [];
+    const batches: Uint8Array[] = [];
+    let finish!: (text: string) => void;
+    const pending = new Promise<string>((resolve) => {
+      finish = resolve;
+    });
+    const transcriber = createWhisperTranscriber({
+      runner: (pcm) => {
+        batches.push(pcm);
+        return pending;
+      },
+      onTranscript: (text) => transcripts.push(text),
+      batchSeconds: 1,
+    });
+
+    transcriber.appendPcm(new Uint8Array(32000).fill(7));
+    transcriber.stop();
+    transcriber.stop();
+    transcriber.appendPcm(new Uint8Array(32000).fill(9));
+    finish('accepted audio');
+    await pending;
+    await Promise.resolve();
+
+    expect(batches).toHaveLength(1);
+    expect(batches[0]).toEqual(new Uint8Array(32000).fill(7));
+    expect(transcripts).toEqual(['accepted audio']);
+  });
+
+  test('delivers both an in-flight batch and the buffered tail on stop', async () => {
+    const transcripts: string[] = [];
+    const completions: ((text: string) => void)[] = [];
+    const pending: Promise<string>[] = [];
+    const transcriber = createWhisperTranscriber({
+      runner: () => {
+        const result = new Promise<string>((resolve) => completions.push(resolve));
+        pending.push(result);
+        return result;
+      },
+      onTranscript: (text) => transcripts.push(text),
+      batchSeconds: 1,
+    });
+
+    transcriber.appendPcm(new Uint8Array(32000));
+    transcriber.appendPcm(new Uint8Array(2));
+    transcriber.stop();
+    completions[0]('full batch');
+    await pending[0];
+    completions[1]('tail');
+    await pending[1];
+    await Promise.resolve();
+
+    expect(transcripts).toEqual(['full batch', 'tail']);
+  });
+
+  test('keeps input order when the tail completes before earlier batches', async () => {
+    const transcripts: string[] = [];
+    const completions: ((text: string) => void)[] = [];
+    const pending: Promise<string>[] = [];
+    const transcriber = createWhisperTranscriber({
+      runner: () => {
+        const result = new Promise<string>((resolve) => completions.push(resolve));
+        pending.push(result);
+        return result;
+      },
+      onTranscript: (text) => transcripts.push(text),
+      batchSeconds: 1,
+    });
+
+    transcriber.appendPcm(new Uint8Array(32000));
+    transcriber.appendPcm(new Uint8Array(32000));
+    transcriber.appendPcm(new Uint8Array(2));
+    transcriber.stop();
+    completions[2]('tail');
+    await pending[2];
+    completions[1]('');
+    await pending[1];
+    expect(transcripts).toEqual([]);
+    completions[0]('first batch');
+    await pending[0];
+    await Promise.resolve();
+    expect(transcripts).toEqual(['first batch', 'tail']);
+  });
 });
 
 describe('parakeetWsUrl', () => {
